+++ cpp_src/core/nsselecter/nsselecter.h
@@ -66,7 +66,7 @@ private:
 									   bool &skipSortingEntry, StrictMode);
 	void getSortIndexValue(const SortingContext &sortCtx, IdType rowId, VariantArray &value, uint8_t proc, const joins::NamespaceResults &,
 						   const JoinedSelectors &);
-	void processLeftJoins(QueryResults &qr, SelectCtx &sctx, size_t startPos);
+	void processLeftJoins(QueryResults &qr, SelectCtx &sctx, size_t startPos, const RdxContext&);
 	bool checkIfThereAreLeftJoins(SelectCtx &sctx) const;
 	template <typename It>
 	void sortResults(LoopCtx &sctx, It begin, It end, const SortingOptions &sortingOptions);