+++ cpp_src/replicator/replicator.cc
@@ -841,8 +841,10 @@ void Replicator::OnWALUpdate(LSNPair LSNs, string_view nsName, const WALRecord &
 		auto replState = slaveNs->GetReplState(dummyCtx_);
 		if (!replState.lastUpstreamLSN.isEmpty()) {
 			if (replState.lastUpstreamLSN >= LSNs.upstreamLSN_) {
-				logPrintf(LogTrace, "[repl:%s:%s]:%d OnWALUpdate old record state = %d upstreamLSN = %s", nsName, slave_->storagePath_,
-						  config_.serverId, state_.load(std::memory_order_acquire), LSNs.upstreamLSN_);
+				logPrintf(LogTrace,
+						  "[repl:%s:%s]:%d OnWALUpdate old record state = %d upstreamLSN = %s replState.lastUpstreamLSN=%d wrec.type = %d",
+						  nsName, slave_->storagePath_, config_.serverId, state_.load(std::memory_order_acquire), LSNs.upstreamLSN_,
+						  replState.lastUpstreamLSN, wrec.type);
 				return;
 			}
 		}